pallet_unified_accounts::pallet::dispatchables

Function claim_evm_address

pub fn claim_evm_address<T: Config>(
    evm_address: EvmAddress,
    signature: [u8; 65],
)

Claim account mapping between Substrate account and Evm address. Ensure no prior mapping exists for evm address.

evm_address: The evm address to bind to the caller’s account

signature: A signature generated by the address to prove ownership

WARNING:

This extrinsic only handles transfer of native balance, if your EVM address contains any other native assets like XC20, DAppStaking unclaimed rewards, etc you need to transfer them before hand, otherwise FUNDS WILL BE LOST FOREVER.

Once connected user cannot change their mapping EVER.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::claim_evm_address.