Type Alias pallet_unified_accounts::pallet::Module

pub type Module<T> = Pallet<T>;

๐Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn claim_evm_address( origin: OriginFor<T>, evm_address: EvmAddress, signature: [u8; 65] ) -> DispatchResult

Claim account mapping between Substrate account and Evm address. Ensure no prior mapping exists for evm address.

evm_address: The evm address to bind to the callerโs account

signature: A signature generated by the address to prove ownership

WARNING:

This extrisic only handles transfer of native balance, if your EVM address contains any other native assets like XC20, DAppStaking unclaimed rewards, etc you need to transfer them before hand, otherwise FUNDS WILL BE LOST FOREVER.

Once connected user cannot change their mapping EVER.

pub fn claim_default_evm_address(origin: OriginFor<T>) -> DispatchResult

Claim default evm address for given account id Ensure no prior mapping exists for the account

WARNINGS: Once connected user cannot change their mapping EVER.

impl<T: Config> Pallet<T>

EIP-712 compatible signature scheme for verifying ownership of EVM Address https://eips.ethereum.org/EIPS/eip-712

Raw Data = Domain Separator + Type Hash + keccak256(AccountId)

pub fn build_signing_payload(who: &T::AccountId) -> [u8; 32]

pub fn verify_signature( who: &T::AccountId, sig: &[u8; 65] ) -> Option<EvmAddress>

impl<T: Config> Pallet<T>

pub fn eth_sign_prehash(prehash: &[u8; 32], secret: &SecretKey) -> [u8; 65]

Sign the given prehash with provided eth private key

pub fn eth_address(secret: &SecretKey) -> EvmAddress

Get the eth address for given eth private key